Separate out home and about pages
Create a home page - this is the landing page, that's the default route.

Place personal info in the about page.
