+++ src/screens/About/About.tsx
@@ -3,8 +3,6 @@ import React from 'react';
 import 'main.css';
 
 import {
-  Typography,
-  TypographyProps,
   styled,
   Box,
   BoxProps,
@@ -13,79 +11,45 @@ import {
 import about1 from "static/images/about_1.png";
 import about2 from "static/images/about_2.png";
 
-import { InstaFeed, AboutCard, MailchimpFormContainer } from 'components';
-
-const WhiteText = styled(Typography)<TypographyProps>(({ theme }) => ({
-  color: "#FFFFFF",
-}));
+import { AboutCard } from 'components';
 
 const Container = styled(Box)<BoxProps>(({ theme }) => ({
   display: 'flex',
   flexDirection: 'column',
   alignItems: 'center',
+  marginTop: 60,
   [theme.breakpoints.up('xs')]: {
-    marginTop: 110,
     width: '100%',
   },
-  [theme.breakpoints.up('sm')]: {
-    marginTop: 160,
-  },
   [theme.breakpoints.up('md')]: {
-    marginTop: 210,
     width: '80%',
     marginLeft: 'auto',
     marginRight: 'auto',
   },
   [theme.breakpoints.up('lg')]: {
-    marginTop: 260,
     width: '60%',
-  },
-}));
-
-const Section = styled(Box)<BoxProps>(({ theme }) => ({
-  display: 'flex',
-  width: '100%',
-  [theme.breakpoints.up('xs')]: {
-    flexDirection: 'column',
-    marginTop: '50px',
-    aligntItems: 'center',
-  },
-  [theme.breakpoints.up('md')]: {
-    flexDirection: 'row',
-    marginTop: '100px',
-    alignItems: 'flex-end',
   },
 }));
 
 const loremIpsum = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
 
 const About = () => {
-  const postUrl = process.env.REACT_APP_MAILCHIMP_URL;
-
   return (
     <Container>
-      <WhiteText variant="h1">Revolutionise your life.</WhiteText>
-      <WhiteText variant="h1">Permanentely.</WhiteText>
       <AboutCard
         imageUrl={about1}
         title="Who am I?"
         subtitle="Subtitle"
         body={loremIpsum}
-        sx={{ mt: 8 }}
+        sx={{ mt: 2 }}
       />
       <AboutCard
         imageUrl={about2}
         title="What can I do for you?"
         subtitle="Subtitle"
         body={loremIpsum}
-        sx={{ mt: 4, mb: 4 }}
+        sx={{ mt: 4, mb: 2 }}
       />
-      <InstaFeed />
-      {postUrl &&
-        <Section marginBottom="40px">
-          <MailchimpFormContainer postUrl={postUrl}/>
-        </Section>
-      }
     </Container>
   )
 }
