+++ src/screens/index.ts
@@ -1,3 +1,4 @@
+export * from "./home";
 export * from "./about";
 export * from "./blog";
 export * from "./shop";
